feat(auth): SSO/MFA security hardening - PKCE S256-only, input validation, rate limiting, PBKDF2 passwords, AES-256-GCM TOTP encryption, QR code generation, automated tests

+++ src/web/http/http_utils.cpp
@@ -94,7 +94,7 @@ namespace httputils {
                 escaped << c;
             } else {
                 escaped << std::uppercase;
-                escaped << '%' << std::setw(2) << static_cast<unsigned char>(c);
+                escaped << '%' << std::setw(2) << static_cast<int>(static_cast<unsigned char>(c));
                 escaped << std::nouppercase;
             }
         }
